+++ CHANGELOG.md
@@ -40,6 +40,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Features
 
 - [#77](https://github.com/alleslabs/celatone-frontend/pull/77) Wireup code info section in code details page
+- [#62](https://github.com/alleslabs/celatone-frontend/pull/62) Add footer
+- [#71](https://github.com/alleslabs/celatone-frontend/pull/71) Add search bar at the top (currently support only contract address and code id)
+- [#69](https://github.com/alleslabs/celatone-frontend/pull/69) Add execute table in contract details page
 - [#68](https://github.com/alleslabs/celatone-frontend/pull/63) Refactor past txs link props and make sure navigation works
 - [#65](https://github.com/alleslabs/celatone-frontend/pull/60) Create instantiate button component
 - [#64](https://github.com/alleslabs/celatone-frontend/pull/64) Add contract not exist page